Move method: Move methods from JWTVerifier to classes implemented from interface
